requirements.txt
streamlit
langchain
openai
tiktoken
chromadb
pypdf
pypdf2
python-docx
docx2txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib